requirements.txt
aiohttp==3.5.4
altgraph==0.16.1
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
discord.py==1.0.1
future==0.17.1
idna==2.8
idna-ssl==1.1.0
jsons==0.8.8
macholib==1.11
multidict==4.5.2
pefile==2019.4.18
PyInstaller==3.4
pywin32-ctypes==0.2.0
requests==2.21.0
typing-extensions==3.7.2
urllib3==1.24.3
websockets==6.0
yarl==1.3.0
